Add support for core packages.
Tests sometimes fail because they can't find packages from the core
images.

+++ packages/fuchsia_ctl/bin/main.dart
@@ -97,7 +97,8 @@ Future<void> main(List<String> args) async {
     ..addMultiOption('far',
         abbr: 'f', help: 'The .far files to include for the test.')
     ..addOption('timeout-seconds',
-        defaultsTo: '120', help: 'Test timeout in seconds.');
+        defaultsTo: '120', help: 'Test timeout in seconds.')
+    ..addOption('packages-directory', help: 'amber files directory.');
 
   final ArgResults results = parser.parse(args);
 
@@ -256,26 +257,34 @@ Future<OperationResult> test(
   ArgResults args,
 ) async {
   const FileSystem fs = LocalFileSystem();
-  final String uuid = Uuid().v4();
   final String identityFile = args['identity-file'];
-  final Directory repo = fs.systemTempDirectory.childDirectory('repo_$uuid');
+
   final PackageServer server = PackageServer(args['pm-path']);
   const SshClient ssh = SshClient();
   final List<String> farFiles = args['far'];
   final String target = args['target'];
   final String arguments = args['arguments'];
+  Directory repo;
+  if (args['packages-directory'] == null) {
+    final String uuid = Uuid().v4();
+    repo = fs.systemTempDirectory.childDirectory('repo_$uuid');
+  } else {
+    repo = fs.directory(args['packages-directory']);
+  }
+
   try {
     final String targetIp = await devFinder.getTargetAddress(deviceName);
     final AmberCtl amberCtl = AmberCtl(targetIp, identityFile);
+    OperationResult result;
     stdout.writeln('Using ${repo.path} as repo to serve to $targetIp...');
-    repo.createSync(recursive: true);
-    OperationResult result = await server.newRepo(repo.path);
-
-    if (!result.success) {
-      stderr.writeln('Failed to create repo at $repo.');
-      return result;
+    if (!await repo.exists()) {
+      repo.createSync(recursive: true);
+      result = await server.newRepo(repo.path);
+      if (!result.success) {
+        stderr.writeln('Failed to create repo at $repo.');
+        return result;
+      }
     }
-
     await server.serveRepo(repo.path, port: 0);
     await amberCtl.addSrc(server.serverPort);
 
@@ -312,7 +321,7 @@ Future<OperationResult> test(
     return testResult;
   } finally {
     // We may not have created the repo if dev finder errored first.
-    if (repo.existsSync()) {
+    if (repo.existsSync() && args['packages-directory']) {
       repo.deleteSync(recursive: true);
     }
     if (server.serving) {
